Rename ToolSpec to Tool throughout the repository (#599)
