+++ examples/18_send_message_while_processing.py
@@ -50,7 +50,7 @@ from openhands.sdk import (
     Agent,
     Conversation,
 )
-from openhands.sdk.tool import ToolSpec, register_tool
+from openhands.sdk.tool import Tool, register_tool
 from openhands.tools.execute_bash import BashTool
 from openhands.tools.str_replace_editor import FileEditorTool
 
@@ -70,10 +70,10 @@ cwd = os.getcwd()
 register_tool("BashTool", BashTool)
 register_tool("FileEditorTool", FileEditorTool)
 tools = [
-    ToolSpec(
+    Tool(
         name="BashTool",
     ),
-    ToolSpec(name="FileEditorTool"),
+    Tool(name="FileEditorTool"),
 ]
 
 # Agent
