+++ examples/25_anthropic_thinking.py
@@ -13,7 +13,7 @@ from openhands.sdk import (
     RedactedThinkingBlock,
     ThinkingBlock,
 )
-from openhands.sdk.tool import ToolSpec, register_tool
+from openhands.sdk.tool import Tool, register_tool
 from openhands.tools.execute_bash import BashTool
 
 
@@ -30,7 +30,7 @@ llm = LLM(
 
 # Setup agent with bash tool
 register_tool("BashTool", BashTool)
-agent = Agent(llm=llm, tools=[ToolSpec(name="BashTool")])
+agent = Agent(llm=llm, tools=[Tool(name="BashTool")])
 
 
 # Callback to display thinking blocks
