+++ openhands/sdk/__init__.py
@@ -36,9 +36,9 @@ from openhands.sdk.mcp import (
 from openhands.sdk.tool import (
     Action,
     Observation,
+    Tool,
     ToolBase,
     ToolDefinition,
-    ToolSpec,
     list_registered_tools,
     register_tool,
     resolve_tool,
@@ -65,9 +65,9 @@ __all__ = [
     "ImageContent",
     "ThinkingBlock",
     "RedactedThinkingBlock",
+    "Tool",
     "ToolDefinition",
     "ToolBase",
-    "ToolSpec",
     "AgentBase",
     "Agent",
     "Action",
