+++ openhands/sdk/tool/__init__.py
@@ -10,7 +10,7 @@ from openhands.sdk.tool.schema import (
     Action,
     Observation,
 )
-from openhands.sdk.tool.spec import ToolSpec
+from openhands.sdk.tool.spec import Tool
 from openhands.sdk.tool.tool import (
     ExecutableTool,
     ToolAnnotations,
@@ -21,9 +21,9 @@ from openhands.sdk.tool.tool import (
 
 
 __all__ = [
+    "Tool",
     "ToolDefinition",
     "ToolBase",
-    "ToolSpec",
     "ToolAnnotations",
     "ToolExecutor",
     "ExecutableTool",
