+++ tests/cross/test_hello_world.py
@@ -23,7 +23,7 @@ from openhands.sdk.event.llm_convertible import (
     MessageEvent,
     ObservationEvent,
 )
-from openhands.sdk.tool import ToolSpec, register_tool
+from openhands.sdk.tool import Tool, register_tool
 from openhands.tools.execute_bash import BashTool
 from openhands.tools.str_replace_editor import FileEditorTool
 
@@ -163,16 +163,16 @@ class TestHelloWorld:
             api_key=SecretStr("mock-api-key"),
         )
 
-        # Tools setup with temporary directory - use registry + ToolSpec as in runtime
+        # Tools setup with temporary directory - use registry + Tool as in runtime
         register_tool("BashTool", BashTool)
         register_tool("FileEditorTool", FileEditorTool)
-        tool_specs = [
-            ToolSpec(name="BashTool"),
-            ToolSpec(name="FileEditorTool"),
+        tools = [
+            Tool(name="BashTool"),
+            Tool(name="FileEditorTool"),
         ]
 
         # Agent setup
-        agent = Agent(llm=llm, tools=tool_specs)
+        agent = Agent(llm=llm, tools=tools)
 
         # Conversation setup
         conversation = Conversation(
@@ -283,16 +283,16 @@ class TestHelloWorld:
             api_key=SecretStr("mock-api-key"),
         )
 
-        # Tools setup with temporary directory - use registry + ToolSpec as in runtime
+        # Tools setup with temporary directory - use registry + Tool as in runtime
         register_tool("BashTool", BashTool)
         register_tool("FileEditorTool", FileEditorTool)
-        tool_specs = [
-            ToolSpec(name="BashTool"),
-            ToolSpec(name="FileEditorTool"),
+        tools = [
+            Tool(name="BashTool"),
+            Tool(name="FileEditorTool"),
         ]
 
         # Create agent and conversation
-        agent = Agent(llm=llm, tools=tool_specs)
+        agent = Agent(llm=llm, tools=tools)
         conversation = Conversation(
             agent=agent,
             workspace=self.temp_dir,
