+++ tests/integration/base.py
@@ -19,7 +19,7 @@ from openhands.sdk.event.base import Event
 from openhands.sdk.event.llm_convertible import (
     MessageEvent,
 )
-from openhands.sdk.tool import ToolSpec
+from openhands.sdk.tool import Tool
 
 
 class TestResult(BaseModel):
@@ -117,7 +117,7 @@ class BaseIntegrationTest(ABC):
 
     @property
     @abstractmethod
-    def tools(self) -> list[ToolSpec]:
+    def tools(self) -> list[Tool]:
         """List of tools available to the agent."""
         pass
