+++ tests/sdk/conversation/local/test_conversation_pause_functionality.py
@@ -32,9 +32,9 @@ from openhands.sdk.security.confirmation_policy import AlwaysConfirm
 from openhands.sdk.tool import (
     Action,
     Observation,
+    Tool,
     ToolDefinition,
     ToolExecutor,
-    ToolSpec,
     register_tool,
 )
 
@@ -106,7 +106,7 @@ class TestPauseFunctionality:
 
         self.agent = Agent(
             llm=self.llm,
-            tools=[ToolSpec(name="test_tool")],
+            tools=[Tool(name="test_tool")],
         )
         self.conversation = Conversation(agent=self.agent)
 
@@ -300,7 +300,7 @@ class TestPauseFunctionality:
         register_tool("test_tool", _make_blocking_tool)
         agent = Agent(
             llm=self.llm,
-            tools=[ToolSpec(name="test_tool")],
+            tools=[Tool(name="test_tool")],
         )
         conversation = Conversation(agent=agent, stuck_detection=False)
