+++ tests/tools/execute_bash/test_conversation_cleanup.py
@@ -9,7 +9,7 @@ import tempfile
 from unittest.mock import Mock
 
 from openhands.sdk import Agent, Conversation
-from openhands.sdk.tool import ToolSpec, register_tool
+from openhands.sdk.tool import Tool, register_tool
 from openhands.tools.execute_bash import BashExecutor, BashTool
 
 
@@ -30,9 +30,7 @@ def test_conversation_close_calls_executor_close(mock_llm):
         # Create agent and conversation
         agent = Agent(
             llm=mock_llm,
-            tools=[
-                ToolSpec(name="test_execute_bash", params={"working_dir": temp_dir})
-            ],
+            tools=[Tool(name="test_execute_bash", params={"working_dir": temp_dir})],
         )
         conversation = Conversation(agent=agent, workspace=temp_dir)
 
@@ -60,9 +58,7 @@ def test_conversation_del_calls_close(mock_llm):
         # Create agent and conversation
         agent = Agent(
             llm=mock_llm,
-            tools=[
-                ToolSpec(name="test_execute_bash", params={"working_dir": temp_dir})
-            ],
+            tools=[Tool(name="test_execute_bash", params={"working_dir": temp_dir})],
         )
         conversation = Conversation(agent=agent, workspace=temp_dir)
 
@@ -93,9 +89,7 @@ def test_conversation_close_handles_executor_exceptions(mock_llm):
         # Create agent and conversation
         agent = Agent(
             llm=mock_llm,
-            tools=[
-                ToolSpec(name="test_execute_bash", params={"working_dir": temp_dir})
-            ],
+            tools=[Tool(name="test_execute_bash", params={"working_dir": temp_dir})],
         )
         conversation = Conversation(agent=agent, workspace=temp_dir)
 
@@ -120,9 +114,7 @@ def test_conversation_close_skips_none_executors(mock_llm):
         # Create agent and conversation
         agent = Agent(
             llm=mock_llm,
-            tools=[
-                ToolSpec(name="test_execute_bash", params={"working_dir": temp_dir})
-            ],
+            tools=[Tool(name="test_execute_bash", params={"working_dir": temp_dir})],
         )
         conversation = Conversation(agent=agent, workspace=temp_dir)
